+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `list` operation to to retrieve entire service catalog
+
+### Changed
+
+- change parameter of `get` and `unregister`
+
 ## [0.0.1] - 24-03-2020
 
 ### Added